fix for no active tab instance
found this issue when hitting refresh on the browser after enabling the disabled tab, and having that be the what should have been the active tab when the page reloaded.  the script hit the error and no tabpanel was shown on reload.   To prevent this from happening, if the active tab cannot be found (for this reason or maybe another?) the first tab in the tablist will be used as the active tab, which seems like that should be a safe fallbck.